+++ src/main/java/com/assetflow/asset/controller/AssetController.java
@@ -1,0 +1,28 @@
+package com.assetflow.asset.controller;
+
+import com.assetflow.asset.Asset;
+import com.assetflow.asset.dto.AssetCreateRequest;
+import com.assetflow.asset.dto.AssetCreateResponse;
+import com.assetflow.asset.service.AssetService;
+import com.assetflow.member.dto.MemberCreateResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/assets")
+public class AssetController {
+    private final AssetService assetService;
+
+    @PostMapping
+    public AssetCreateResponse AssetCreate(@RequestBody AssetCreateRequest request) {
+        return assetService.createAsset(request);
+    }
+
+    @DeleteMapping("/delete/{assetId}")
+    public ResponseEntity<Void> assetDelete(@PathVariable("assetId") Long assetId ) {
+        assetService.assetDelete(assetId);
+        return ResponseEntity.noContent().build();
+    }
+}