+++ src/main/java/com/assetflow/asset/dto/AssetCreateRequest.java
@@ -7,8 +7,9 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class AssetCreateRequest {
+    private Long
     private String name;
     private String explanation;
     private AssetType assetType;
-    private Long category;
+    private Long categoryId;
 }