+++ src/main/java/com/assetflow/asset/dto/AssetItemCreateResponse.java
@@ -1,0 +1,16 @@
+package com.assetflow.asset.dto;
+
+import com.assetflow.asset.Asset;
+import com.assetflow.asset.AssetItemStatus;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class AssetItemCreateResponse {
+    private Long assetItemId;
+    private String serialNumber;
+    private String location;
+    private Long assetId;
+    private AssetItemStatus assetItemStatus;
+}