+++ src/main/java/com/assetflow/asset/dto/CategoryCreateResponse.java
@@ -1,0 +1,11 @@
+package com.assetflow.asset.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CategoryCreateResponse {
+    private Long id;
+    private String name;
+}