+++ src/main/java/com/assetflow/global/exception/ServiceExceptionHandler.java
@@ -1,0 +1,18 @@
+package com.assetflow.global.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ServiceExceptionHandler {
+    @ExceptionHandler(IllegalStateException.class)
+    public ResponseEntity<ErrorResult> illegalExHandle(IllegalStateException e) {
+        ErrorResult errorResponse = new ErrorResult("BAD_REQUEST", e.getMessage());
+
+        return ResponseEntity
+                .status(HttpStatus.BAD_REQUEST)
+                .body(errorResponse);
+    }
+}